+++ lino_welfare/modlib/client_vouchers/__init__.py
@@ -40,6 +40,6 @@ class Plugin(ad.Plugin):
     verbose_name = _("Client vouchers")
 
     def setup_explorer_menu(self, site, user_type, m):
-        mg = site.plugins.accounts
+        mg = site.plugins.ledger
         m = m.add_menu(mg.app_label, mg.verbose_name)
         m.add_action('client_vouchers.Voucher')