+++ lino_welfare/modlib/debts/fixtures/minimal.py
@@ -29,9 +29,6 @@ from lino.utils.instantiator import Instantiator
 def objects():
     AccountTypes = rt.models.debts.AccountTypes
     TableLayouts = rt.models.debts.TableLayouts
-    # c = accounts.Chart(name="debts.default")
-    # yield c
-    # c = rt.models.accounts.AccountCharts.debts
     group = Instantiator('debts.Group').build
     g = group(
         ref="10", account_type=AccountTypes.incomes,